Extract ExternalPadExportFetcher out of EtherpadClient
Closes #62.

EtherpadClient was 683 LOC and mixed three concerns: the Etherpad
admin-API client, external public-pad export fetching, and the
SSRF-hardened HTTP layer that only the external path needs. The
admin-API surface dragged the SSRF / IP-pinning code into every
service that just wanted to talk to Etherpad.

Carved out:

- ExternalPadExportFetcher (new, 358 LOC) — owns
  normalizeAndFetchExternalPublicPadText{,OrEmpty},
  assertPublicPadAvailable, normalizeAndValidateExternalPublicPadUrl,
  getPublicTextFromPadUrl and the resolve/parse/SSRF-pinning helpers
  (parsePublicPadUrl, resolveAndValidateExternalHost, isPublicIp,
  isAllowlistedExternalHost, normalizeAllowlistOrigin,
  buildPublicExportUrl, sendPinnedPublicGetRequest,
  assertAllowedExternalExportContentType,
  getPublicTextFromResolvedExternalPad,
  resolveAndValidateExternalPublicPadUrl).
- EtherpadClient (slimmed to 347 LOC) — keeps the admin-API methods,
  healthCheck, detectApiVersion, plus the host/origin accessors
  (getConfiguredOrigin, normalizeOrigin, getPublicHost) and the
  shared apiCall / sendRequest / sendPublicGetRequest plumbing.

Callers updated:

- ExternalPadSeeder now depends on ExternalPadExportFetcher only
  (dropped EtherpadClient dep entirely).
- PadSyncService, PadOpenService, PadMetadataService,
  PublicPadOpenService inject the fetcher in addition to
  EtherpadClient for the external-pad path.

Tests:

- New ExternalPadExportFetcherTest covers the moved external-URL
  validation + fetch tests.
- EtherpadClientTest slimmed to admin-API + origin helpers.
- 8 service tests adjusted to mock the fetcher where they
  previously stubbed external methods on EtherpadClient.

395 tests stay green (one less than before because the moved tests
re-organize rather than multiply).
