+++ lib/Service/ExternalPadExportFetcher.php
@@ -1,0 +1,358 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ * Copyright (c) 2026 Jacob Bühler
+ */
+
+namespace OCA\EtherpadNextcloud\Service;
+
+use OCA\EtherpadNextcloud\Exception\EtherpadClientException;
+use OCA\EtherpadNextcloud\Exception\ExternalPadExportNotFoundException;
+use OCP\IConfig;
+
+class ExternalPadExportFetcher {
+	private const EXTERNAL_EXPORT_MAX_BYTES = 5242880; // 5 MiB
+	private const EXTERNAL_REQUEST_TIMEOUT_SECONDS = 15;
+
+	public function __construct(
+		private IConfig $config,
+	) {
+	}
+
+	public function getPublicTextFromPadUrl(string $padUrl): string {
+		$resolved = $this->resolveAndValidateExternalPublicPadUrl($padUrl);
+		return $this->getPublicTextFromResolvedExternalPad($resolved);
+	}
+
+	/** @return array{origin:string,pad_id:string,pad_url:string,text:string} */
+	public function normalizeAndFetchExternalPublicPadText(string $padUrl): array {
+		$resolved = $this->resolveAndValidateExternalPublicPadUrl($padUrl);
+		return [
+			'origin' => $resolved['origin'],
+			'pad_id' => $resolved['pad_id'],
+			'pad_url' => $resolved['pad_url'],
+			'text' => $this->getPublicTextFromResolvedExternalPad($resolved),
+		];
+	}
+
+	/**
+	 * @return array{origin:string,pad_id:string,pad_url:string,text:string,snapshot_unavailable:bool}
+	 */
+	public function normalizeAndFetchExternalPublicPadTextOrEmpty(string $padUrl): array {
+		$resolved = $this->resolveAndValidateExternalPublicPadUrl($padUrl);
+		$snapshotUnavailable = false;
+		try {
+			$text = $this->getPublicTextFromResolvedExternalPad($resolved);
+		} catch (ExternalPadExportNotFoundException) {
+			$text = '';
+			$snapshotUnavailable = true;
+		}
+
+		return [
+			'origin' => $resolved['origin'],
+			'pad_id' => $resolved['pad_id'],
+			'pad_url' => $resolved['pad_url'],
+			'text' => $text,
+			'snapshot_unavailable' => $snapshotUnavailable,
+		];
+	}
+
+	public function assertPublicPadAvailable(string $padUrl): void {
+		$this->getPublicTextFromPadUrl($padUrl);
+	}
+
+	/** @return array{origin:string,pad_id:string,pad_url:string} */
+	public function normalizeAndValidateExternalPublicPadUrl(string $padUrl): array {
+		$resolved = $this->resolveAndValidateExternalPublicPadUrl($padUrl);
+		return [
+			'origin' => $resolved['origin'],
+			'pad_id' => $resolved['pad_id'],
+			'pad_url' => $resolved['pad_url'],
+		];
+	}
+
+	private function buildPublicExportUrl(string $padUrl, string $format): string {
+		$parsed = $this->parsePublicPadUrl($padUrl);
+		return $parsed['pad_url'] . '/export/' . $format;
+	}
+
+	/**
+	 * @param array{pad_url:string,host:string,port:int,resolved_ips:list<string>} $resolved
+	 */
+	private function getPublicTextFromResolvedExternalPad(array $resolved): string {
+		$url = $this->buildPublicExportUrl($resolved['pad_url'], 'txt');
+		return $this->sendPinnedPublicGetRequest($url, $resolved['host'], $resolved['port'], $resolved['resolved_ips']);
+	}
+
+	/** @return array{origin:string,pad_id:string,pad_url:string,host:string,port:int,resolved_ips:list<string>} */
+	private function resolveAndValidateExternalPublicPadUrl(string $padUrl): array {
+		$parsed = $this->parsePublicPadUrl($padUrl);
+		$padId = $parsed['pad_id'];
+		if (preg_match('/^g\.[^$]+\$.+$/', $padId) === 1) {
+			throw new EtherpadClientException('Only public pad URLs can be linked from external servers.');
+		}
+
+		return [
+			'origin' => $parsed['origin'],
+			'pad_id' => $parsed['pad_id'],
+			'pad_url' => $parsed['pad_url'],
+			'host' => $parsed['host'],
+			'port' => $parsed['port'],
+			'resolved_ips' => $this->resolveAndValidateExternalHost($parsed['host'], $parsed['origin']),
+		];
+	}
+
+	/**
+	 * @param list<string> $resolvedIps
+	 */
+	private function sendPinnedPublicGetRequest(string $url, string $host, int $port, array $resolvedIps): string {
+		if (!function_exists('curl_init')) {
+			throw new EtherpadClientException('External pad sync requires PHP cURL extension.');
+		}
+
+		$errors = [];
+		foreach ($resolvedIps as $ip) {
+			$buffer = '';
+			$contentType = '';
+			$sizeExceeded = false;
+			$maxBytes = self::EXTERNAL_EXPORT_MAX_BYTES;
+			$curl = curl_init($url);
+			if ($curl === false) {
+				throw new EtherpadClientException('Could not initialize external export request.');
+			}
+			$curlOptions = [
+				CURLOPT_RETURNTRANSFER => false,
+				CURLOPT_FOLLOWLOCATION => false,
+				CURLOPT_MAXREDIRS => 0,
+				CURLOPT_CONNECTTIMEOUT => self::EXTERNAL_REQUEST_TIMEOUT_SECONDS,
+				CURLOPT_TIMEOUT => self::EXTERNAL_REQUEST_TIMEOUT_SECONDS,
+				CURLOPT_HTTPGET => true,
+				CURLOPT_SSL_VERIFYPEER => true,
+				CURLOPT_SSL_VERIFYHOST => 2,
+				CURLOPT_HTTPHEADER => [
+					'Accept: text/plain, application/octet-stream;q=0.9, */*;q=0.1',
+				],
+				CURLOPT_RESOLVE => [$host . ':' . $port . ':' . $ip],
+				CURLOPT_HEADERFUNCTION => static function ($ch, string $headerLine) use (&$contentType): int {
+					$line = trim($headerLine);
+					if ($line !== '' && stripos($line, 'Content-Type:') === 0) {
+						$contentType = trim((string)substr($line, strlen('Content-Type:')));
+					}
+					return strlen($headerLine);
+				},
+				CURLOPT_WRITEFUNCTION => static function ($ch, string $chunk) use (&$buffer, &$sizeExceeded, $maxBytes): int {
+					if (strlen($buffer) + strlen($chunk) > $maxBytes) {
+						$sizeExceeded = true;
+						return 0;
+					}
+					$buffer .= $chunk;
+					return strlen($chunk);
+				},
+			];
+			if (defined('CURLOPT_PROTOCOLS') && defined('CURLPROTO_HTTPS')) {
+				$curlOptions[CURLOPT_PROTOCOLS] = CURLPROTO_HTTPS;
+			}
+			if (defined('CURLOPT_REDIR_PROTOCOLS') && defined('CURLPROTO_HTTPS')) {
+				$curlOptions[CURLOPT_REDIR_PROTOCOLS] = CURLPROTO_HTTPS;
+			}
+			curl_setopt_array($curl, $curlOptions);
+
+			$success = curl_exec($curl);
+			$httpCode = (int)curl_getinfo($curl, CURLINFO_RESPONSE_CODE);
+			$curlError = curl_error($curl);
+			curl_close($curl);
+
+			if ($success === false) {
+				if ($sizeExceeded) {
+					throw new EtherpadClientException(
+						'External public pad export exceeds maximum size (' . self::EXTERNAL_EXPORT_MAX_BYTES . ' bytes).'
+					);
+				}
+				$errors[] = 'transport via ' . $ip . ': ' . ($curlError !== '' ? $curlError : 'unknown error');
+				continue;
+			}
+			if ($httpCode >= 400) {
+				if ($httpCode === 404) {
+					throw new ExternalPadExportNotFoundException(
+						'External public pad export was not found. Make sure the pad exists and can be exported.'
+					);
+				}
+				throw new EtherpadClientException('Public export HTTP error (' . $httpCode . ')');
+			}
+
+			$this->assertAllowedExternalExportContentType($contentType);
+			return $buffer;
+		}
+
+		$detail = $errors !== [] ? implode('; ', $errors) : 'all resolved targets failed';
+		throw new EtherpadClientException('Public export transport error: ' . $detail);
+	}
+
+	private function assertAllowedExternalExportContentType(string $contentTypeHeader): void {
+		$raw = trim($contentTypeHeader);
+		if ($raw === '') {
+			throw new EtherpadClientException('Public export did not provide a Content-Type header.');
+		}
+
+		$normalized = strtolower(trim((string)explode(';', $raw, 2)[0]));
+		if ($normalized === 'text/html') {
+			throw new EtherpadClientException('Public export returned unsupported Content-Type: text/html');
+		}
+		if (str_starts_with($normalized, 'text/')) {
+			return;
+		}
+		if (in_array($normalized, ['application/octet-stream', 'application/text'], true)) {
+			return;
+		}
+
+		throw new EtherpadClientException('Public export returned unsupported Content-Type: ' . $normalized);
+	}
+
+	/** @return list<string> */
+	private function resolveAndValidateExternalHost(string $host, string $origin): array {
+		if ((string)$this->config->getAppValue('etherpad_nextcloud', 'allow_external_pads', 'no') !== 'yes') {
+			throw new EtherpadClientException('External pad linking is disabled by admin settings.');
+		}
+		if (!$this->isAllowlistedExternalHost($host, $origin)) {
+			throw new EtherpadClientException('External pad host is not in the allowlist.');
+		}
+		if ($host === 'localhost' || str_ends_with($host, '.localhost') || str_ends_with($host, '.local')) {
+			throw new EtherpadClientException('Local hosts are not allowed for external pad sync.');
+		}
+
+		if (filter_var($host, FILTER_VALIDATE_IP) !== false) {
+			if (!$this->isPublicIp($host)) {
+				throw new EtherpadClientException('Private/reserved IPs are not allowed for external pad sync.');
+			}
+			return [$host];
+		}
+
+		$records = @dns_get_record($host, DNS_A + DNS_AAAA);
+		if (!is_array($records) || $records === []) {
+			throw new EtherpadClientException('Could not resolve external pad host.');
+		}
+
+		$resolvedIps = [];
+		foreach ($records as $record) {
+			if (isset($record['ip']) && is_string($record['ip']) && $record['ip'] !== '') {
+				$resolvedIps[] = $record['ip'];
+			}
+			if (isset($record['ipv6']) && is_string($record['ipv6']) && $record['ipv6'] !== '') {
+				$resolvedIps[] = $record['ipv6'];
+			}
+		}
+		if ($resolvedIps === []) {
+			throw new EtherpadClientException('Could not resolve external pad host to IP.');
+		}
+
+		foreach ($resolvedIps as $ip) {
+			if (!$this->isPublicIp($ip)) {
+				throw new EtherpadClientException('Private/reserved IPs are not allowed for external pad sync.');
+			}
+		}
+
+		return array_values(array_unique($resolvedIps));
+	}
+
+	private function isPublicIp(string $ip): bool {
+		return filter_var($ip, FILTER_VALIDATE_IP, FILTER_FLAG_NO_PRIV_RANGE | FILTER_FLAG_NO_RES_RANGE) !== false;
+	}
+
+	private function isAllowlistedExternalHost(string $host, string $origin): bool {
+		$raw = trim((string)$this->config->getAppValue('etherpad_nextcloud', 'external_pad_allowlist', ''));
+		if ($raw === '') {
+			return true;
+		}
+
+		$entries = preg_split('/[\s,;]+/', $raw) ?: [];
+		$hostLower = strtolower($host);
+		$originLower = strtolower($origin);
+		foreach ($entries as $entry) {
+			$normalized = strtolower(trim($entry));
+			if ($normalized === '') {
+				continue;
+			}
+			if (preg_match('#^https?://#i', $normalized) === 1) {
+				if ($this->normalizeAllowlistOrigin($normalized) === $originLower) {
+					return true;
+				}
+				continue;
+			}
+			if (trim($normalized, ". \t\n\r\0\x0B") === $hostLower) {
+				return true;
+			}
+		}
+
+		return false;
+	}
+
+	private function normalizeAllowlistOrigin(string $entry): string {
+		$parts = parse_url($entry);
+		if (!is_array($parts)) {
+			return '';
+		}
+		$scheme = strtolower((string)($parts['scheme'] ?? ''));
+		$host = strtolower((string)($parts['host'] ?? ''));
+		$port = isset($parts['port']) ? (int)$parts['port'] : 443;
+		if ($scheme !== 'https' || $host === '' || $port <= 0 || $port > 65535) {
+			return '';
+		}
+		return $port === 443 ? 'https://' . $host : 'https://' . $host . ':' . $port;
+	}
+
+	/** @return array{origin:string,pad_id:string,pad_url:string,host:string,port:int} */
+	private function parsePublicPadUrl(string $padUrl): array {
+		$trimmed = trim($padUrl);
+		if ($trimmed === '' || preg_match('#^https?://#i', $trimmed) !== 1) {
+			throw new EtherpadClientException('Invalid public pad URL.');
+		}
+
+		$parts = parse_url($trimmed);
+		if (!is_array($parts)) {
+			throw new EtherpadClientException('Invalid public pad URL.');
+		}
+		if (isset($parts['user']) || isset($parts['pass'])) {
+			throw new EtherpadClientException('Public pad URL must not contain credentials.');
+		}
+
+		$scheme = strtolower((string)($parts['scheme'] ?? ''));
+		$host = strtolower((string)($parts['host'] ?? ''));
+		$port = isset($parts['port']) ? (int)$parts['port'] : 443;
+		$path = (string)($parts['path'] ?? '');
+		// `+` is literal in URL path segments — only query/form bodies treat
+		// it as a space. Using urldecode here turned `/p/team+pad` into
+		// pad-id `team pad`, then re-encoded to `/p/team%20pad` at fetch
+		// time, hitting a different / non-existent remote pad.
+		$decodedPath = rawurldecode($path);
+		if ($scheme !== 'https' || $host === '' || $decodedPath === '' || $port <= 0 || $port > 65535) {
+			throw new EtherpadClientException('Invalid public pad URL.');
+		}
+
+		if (preg_match('~^(.*)/p/([^/]+)$~', $decodedPath, $matches) !== 1) {
+			throw new EtherpadClientException('Public pad URL must match /p/{padId}.');
+		}
+
+		$basePath = rtrim((string)$matches[1], '/');
+		$padId = trim((string)$matches[2]);
+		if ($padId === '') {
+			throw new EtherpadClientException('Invalid public pad URL.');
+		}
+
+		$origin = $scheme . '://' . $host;
+		if ($port !== 443) {
+			$origin .= ':' . $port;
+		}
+		$normalizedBasePath = $basePath === '' ? '' : $basePath;
+		$canonicalPadUrl = $origin . $normalizedBasePath . '/p/' . rawurlencode($padId);
+
+		return [
+			'origin' => $origin,
+			'pad_id' => $padId,
+			'pad_url' => $canonicalPadUrl,
+			'host' => $host,
+			'port' => $port,
+		];
+	}
+}
